% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcCooc.R
\name{calcCooc}
\alias{calcCooc}
\title{Calculate coccurrence matrix}
\usage{
calcCooc(
  exb,
  relative = TRUE,
  format = "data.frame",
  Variable = "V",
  Variante = "Variante"
)
}
\arguments{
\item{Variante}{}
}
\value{

}
\description{
Calculate coccurrence matrix
}
