% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountPerIP.R
\name{CountPerIP}
\alias{CountPerIP}
\title{Count variants per intonation phrase}
\usage{
CountPerIP(exb, Variable = "V", Variante = "Variante", format = "dataframe")
}
\arguments{
\item{format}{}
}
\value{

}
\description{
Count variants per intonation phrase
}
