% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countVars.R
\name{countVars}
\alias{countVars}
\title{countVars Counting Variables in a EXB file}
\usage{
countVars(Corpus, Variable = "Variable", Variante = "Variante")
}
\arguments{
\item{Variante}{}
}
\value{

}
\description{
countVars Counting Variables in a EXB file
}
