% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCooc.R
\name{plotCooC}
\alias{plotCooC}
\title{Title}
\usage{
plotCooC(
  Corpus,
  Variable = "Variable",
  Variante = "Variante",
  threshold = 10,
  umap.config = umap.defaults
)
}
\arguments{
\item{umap.config}{}
}
\value{

}
\description{
Title
}
